duck-dns-client-back/requirements.txt
APScheduler==3.10.4
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
flask==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tzlocal==5.1
urllib3==2.0.7
werkzeug==3.0.0
zipp==3.17.0
